A patch from Sascah Rodekamp "Server Hit Bin, counterHit receive delegator" (https://issues.apache.org/jira/browse/https://issues.apache.org/jira/browse/OFBIZ-3984) - OFBIZ-3984
Patch to improve the delegator call in the countHit method.
The method got the delegator name from the session and got, with this delegator name over the delegator factory, the delegator.
But when the delegator name was not available in the session the method crashes.
I try to get the delegator from the request, and if their is no delegator object I go the factory way.

I think it's a little bit saver to get a delegator object this way

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1021344 13f79535-47bb-0310-9956-ffa450edef68