Implements the description in lookup which so far was missed in migration.
It works well in all browsers; though I had to add a String.trim() function because IE8 does not implement trim yet. BWT, it was the 1st time I used the integrated IE8 debugger and crossed an issue to have it to work :(.
Also
* In Chrome the description is not at right of the lookup but below. 
* When loading and the field has no value a small yellow colored square appears at right of the lookup (all browsers), should not be a bid deal to fix later...

Also adds a new feature which allows to set a dependent field using autocomplete in lookups (was only working with lookup, not with autocomplete). Convenient when, for instance, you want to allow you users to handle names instead of Ids but need to set an Id key in a hidden field. Demonstrated with Geo in example/control/FormWidgetExamples.
Note: for this new feature at some point I wondered if I should not introduce a class plugin. But finally decided to go my own way with prototypes (actually following jQuery spirit, javascript and prototype are powerful, just another way of thinking about classes)

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/jquery@1028261 13f79535-47bb-0310-9956-ffa450edef68