2 slightly modified patches from Sascha Rodekamp "Content Navigation Tree implemented with jQuery" (https://issues.apache.org/jira/browse/OFBIZ-3815) - OFBIZ-3815 + the elrteEditor jQuery plugin
Sascha: 
Migrates the last WebCms Content Screens.
Adds the jsTree for the CMS Web Site Page and replace the old dojo WYSIWYG Editor by editor from elrte http://elrte.org/.

There is still one issue regarding the new editor. When you write a text and store it with the Editor everything is fine, the HTML code is stored in the electronicText Table.
But when you reload the site and the editor get the stored data from the DB it has HTML tags in it. Freemarker encode the Tags in HTML Save Characters (works as designed). So we need a method which decodes the HTML Save output into normal HTML output to let the editor handle the data the right way. So we need something like 'encodeXmlValue' (the other way round).

JLR remarks: 
* I have used jQuery 1.4.2 as I just reverted to this version
* I have replaced PageTitleNavigate by PageTitleNavigateDataResources 
* We will need to localize labels and alert messages
    I noticed
    +    function destroyEditor() {
    +        // TODO
    +    }
    not sure it needs something or not
* I have used 'maxi' as elrte Editor option to have table buttons by default
* To decodes HTML as input to elrte Editor, I have tried to use simple Freemarker replaces. I was surprised to see that some were not decoded. I then tried to use a regexp but without more success... to be continued...



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/jquery@1036195 13f79535-47bb-0310-9956-ffa450edef68