Related to OFBIZ-3933
This to prevent issues and set layer as default if someone calls ConstructLookup directly (ie not by using a widget form lookup or using <@htmlTemplate.lookupField in a ftl template. This should not be but I crossed an issue due to this kind of thing in a custom app...). Can't hurt anyway...

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/jquery@1036662 13f79535-47bb-0310-9956-ffa450edef68