A patch from Sascha Rodekamp "Redirect from Http to Https causes double service call" (https://issues.apache.org/jira/browse/OFBIZ-3988) - OFBIZ-3988
The problem is, when you call a request over http, but the controller wants an https request, the RequestHandler make an redirect to https (which is indeed pretty cool). But this redirect causes that Events/ Services will be called twice. This can be avoided when we return from the Service/ Event directly after doing the redirect.
Example: http://localhost:8080/ecommerce/control/setdistributor?distributor=4711

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1040044 13f79535-47bb-0310-9956-ffa450edef68