Bug Fix "DateTimePicker shows wrong current time" (https://issues.apache.org/jira/browse/OFBIZ-4169) - When changing the plugin configuration the dtp shows the correct time - Thanks Nikkel K for reporting.
Orig. Msg:
The DateTimePicker calculates nearly always in 5m steps and rounds up or down, depending on the nearest result. This means, that you have at 14:11:10 and at 14:12:10 a time of 14:10:10 in the DTP and at 14:13:10 and 14:14:10 already 14:15:10 (so on the one hand you have a "current" time that shows in the future and on the other hand the same time 5 times in the DTP (while the mins. dont increase, the seconds do).

comment to the nearly always: i had exactly one case, where i had xx:01:xx, and i cant recreate it...


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1071328 13f79535-47bb-0310-9956-ffa450edef68