A fix suggested by Michael Reichenbach at "Multiple responses send by SOAPEventHandler when ServiceValidationException occurs"   https://issues.apache.org/jira/browse/OFBIZ-4207
Problem
=======

When calling a service that was defined as "export=true" and a non-optional parameter is missing, then the SOAP-Response is not a valid XML.

Cause
=====

The cause is, that in SOAPEventHandler.invoke() the dispatcher.runSync() throws a ServiceValidationException ("The following required parameter is missing..."). This will be catched by the GenericServiceException-block, a error response will be written to the output stream (via sendError()) and an EventHandlerException will be thrown. This exception will again be catched (this time by the Exception-block and another error response will be written to the output stream.

Solution
========

I think the best solution is simply not to call sendError in the GenericServiceException-block.

JLR: I simply commented out the faulty line (with a reference to OFBIZ-4207) in case someone would want to improve the error message.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1075952 13f79535-47bb-0310-9956-ffa450edef68