A patch from Nicolas Malin "Enhanced way of displaying "missing/incorrect fields"" https://issues.apache.org/jira/browse/OFBIZ-2132
Initial demand by Michael Pätzold
At the moment we concentrate on increasing the usability of the backend application. One thing we recognized is the way form errors are reported to the users. Now it's done by showing a list of entity fields, that are required. A more common approach would be to highlight the form controls, that were not/wrongly filled. Was/Is there any task related to that or something similar?

Nicolas's solution
With integration of JQuery and validate function integrate by Sascha, I extend the Form renderer to apply required class and alert the form when she submitted to pass by JQuery validate function.

JLR: I have added a required field to InternalName field in Edit Product Form. There are maybe other cases like that which should be improved...

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1096293 13f79535-47bb-0310-9956-ffa450edef68