FIX/OPTIMIZE: Do column-name munging once, and save the output, instead of doing it each time the sql string is generated.  This solidifies the definition of getColValue() and getColName().  colName is the identifier used to refer to the column, it's generally what comes after the AS keyword.  colValue is used to the left of AS, and in CONDITION/ORDER BY/GROUP BY clauses.
This change should allow EntityFieldValue to be rewritten, so it doesn't
have the hack of having to deal with the entityAliasStack.  In theory,
it can just call getColName/getColValue(I'm not certain which is the
correct one for this case).

Also, EntityFieldValue should not have that hack at all, it should have
been put into a completely separate class.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1139919 13f79535-47bb-0310-9956-ffa450edef68