A patch from Carsten Schinzer https://issues.apache.org/jira/browse/OFBIZ-4208 "Product page meta-tag does not contain i18n content"
The groovy script that produces the content meta information is only partially using i18n information from the Content Wrapper. Namely, it does NOT use the Product or Category ContentWrapper to receive internationalized content for Product Name and Description or Category Description but rather the direct properties of the product entity. Using the ContentWrapper will overcome this shortcoming.

It is assumed that ContentWrappers will handle primitive over related Content properly, ie. if there is no related Content entity but only primitive content, it will return that. And if there is i18n content missing in the desired (current session) locale, it will return the default locale translation.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1194958 13f79535-47bb-0310-9956-ffa450edef68