A patch from Kiran Gawde "redundant LookupContactList in product app" https://issues.apache.org/jira/browse/OFBIZ-4574
Following lines from product\controller.xml are redundant. They are present in partymgr.

<request-map uri="LookupContactList"><security https="true" auth="true"/><response name="success" type="view" value="LookupContactList"/></request-map>
<view-map name="LookupContactList" type="screen" page="component://party/widget/partymgr/PartyContactListScreens.xml#ListLookupContactList"/>

jleroux: indeed, I don't see any use at all of LookupContactList in product component (only in Marketing and Product). On the other hand, since there are not the same in Marketing and Product it could be seen as ambiguous, and then if someone needs it in catalog it will have to make the choice, maybe the reason it's there?




git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1206693 13f79535-47bb-0310-9956-ffa450edef68