Example Application Is Badly Broken (https://issues.apache.org/jira/browse/OFBIZ-4619). A Bug reported by Adrian Crum: Go to https://localhost:8443/example/control/main
Click the New Example button. In the popup window, click Create. You are taken to the Find Example screen and it contains an error message.

Look at the current URL. Try navigating to it: https://localhost:8443/example/control/createExample The Find Example screen is displayed with an error message.

Click your browser's Back button. The popup dialog appears again.

In summary: There is no way to correct an invalid entry.

I added a required flag for the mandatory fields. 


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1230474 13f79535-47bb-0310-9956-ffa450edef68