A patch from BREMA Dev Team "email password broken" https://issues.apache.org/jira/browse/OFBIZ-3790
There is no productStoreId in the session for an user of a backend application, and ProductStoreEmailSetting is not available. This is the reason.
However we can use default template and default sender's address from general.properties in this case.
Also removes some dead code.

jleroux: this uses  "general.properties", "defaultFromEmailAddress". So to work correctly it still needs "defaultFromEmailAddress" (default =ofbiztest@example.com), "mail.smtp.relay.host" (default =localhost) and such to be set correctly... In other words the message "A new password has been created and sent to you. Please check your Email." could even more misleading if people are not aware about setting SMTP


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1307895 13f79535-47bb-0310-9956-ffa450edef68