A patch from Deepak Dixit for "Auto-completer request should be async." https://issues.apache.org/jira/browse/OFBIZ-4780
Fixes a minor issue regarding the pending request count for ajax auto-completer: pending count is used to add and remove the ui-autocomplete-loading class. If we are aborting the request then this count is not updated properly and ui-autocomplete-loading class does not remove from the lookup element hence ajax loader image display continuously if response came.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1310266 13f79535-47bb-0310-9956-ffa450edef68