A patch from Tom Burns for "Update Company Contact Information Fails" https://issues.apache.org/jira/browse/OFBIZ-4804
Updating the Company postal_address in Party Manager fails.
To reproduce:
In Party Manager lookup Party Id "Company"
Click Update in Contact Information > Postal Address
Make a change to the address
Click Save
Note the error message.

A postal_address update triggers an accounting SECA that runs:
PaymentMethodServices.xml (accounting) updatePaymentMethodAddress
services_paymentmethod.xml updateEftAccount
PaymentMethodServices.java updateEftAccount

The purpose of the SECA is to update the contact_mech_id field in CreditCard or EftAccount entities have the optional billing address information.

The updateEftAccount service sets the nameOnAccount field to optional="false".
The demo data that set up two EftAccount's for Company did not include nameOnAccount.
These are the only ETF's set up in demo data.

In the application ETF's are set up in eCommerce which requires "Name On Account" so the problem should not happen with address changes.
See eCommerce:
Login as DemoCustomer
Click "Profile" at the right of the top title bar
Click "Create New EFT Account" at the right of the title bar in the Payment Method Information section.
Note the first file "Name on Account" is required
Note the option to select a billing address.

Changing the DemoFinAccountData.xml file to enter the nameOnAccount field for the Company EftAccount's fixes the problem.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1328946 13f79535-47bb-0310-9956-ffa450edef68