Overhauled Mini-language <check-errors> and <add-error> elements.
This commit fixes a bug in the <add-error> element. Some of the tests written in Mini-language fail now because errors were previously ignored, but now the error handling works properly.

To assist the developer community with fixing the tests, I added an INFO log message to track down the faulty test code. Run the tests, and then search the log file for "<add-error>".


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1331802 13f79535-47bb-0310-9956-ffa450edef68