A patch from Nicolas Malin for "merge three maps with map-to-map" https://issues.apache.org/jira/browse/OFBIZ-5152
To merge two or more maps on one map with <map-to-map> element on mini-lang doesn't work.
example :
      <call-service service-name="accReport-header" in-map-name="headerMap">
        <result-to-field result-name="jrParameters" />
      </call-service>
      <map-to-map map="reportParameters" to-map="jrParameters"/>
      <map-to-map map="parameters" to-map="jrParameters"/>

This doesn't work because the map-to-map method doesn't accept to merge an existing to-map and the <map-to-map map="parameters" to-map="jrParameters"/> has been ignored 

The patch move the map merge function to run it when a map is given.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1470151 13f79535-47bb-0310-9956-ffa450edef68