Reverts r1487997 and uses Stephen's patch for Fixes "CheckOutEvents setCheckOutPages" https://issues.apache.org/jira/browse/OFBIZ-4161
There was an extra check in CheckOutEvents.setCheckOutPages that prevented an error message from being produced when no payment method was selected on the "How shall you pay" step of the checkout process. 

jleroux: and you got redirected to the order entry page, with the patch you stay on the payment method page.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1488000 13f79535-47bb-0310-9956-ffa450edef68