A patch from Arpit Singh Pandya for "Cancel/Done is not working correctly on Create User Login page." https://issues.apache.org/jira/browse/OFBIZ-5252
When we navigate on create user login page from party profile on clicking "Create New" link in User Name(s) Section, "Cancel/Done" is not working correctly on Create User Login page. partyId is missing with link. It is not working on trunk and 12.04, not sure about other versions. 

jleroux: 11.04 is concerned, 10.04 is OK

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1497839 13f79535-47bb-0310-9956-ffa450edef68