Simplified the layout of the main/framework/applications/specialpurpose build files; now the files:
framework/build.xml
applications/build.xml
specialpurpose/build.xml

are only used to define the filelist properties (as before) with the list of components to be built.
The main build file simply refers to these lists to build/doc/clean OFBiz; as before, if applications' or specialpurpose's build.xml files are absent, the folders are ignored.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1520744 13f79535-47bb-0310-9956-ffa450edef68