A patch from  Christian Carlow for "Additional ship group renderDateFields not created" https://issues.apache.org/jira/browse/OFBIZ-OFBIZ-5382 - OFBIZ-5382
ShipBeforeDate/ShipAfterDate fields are not created as renderDateFields for additional ship groups on orderview page.

The problem is that additional ship groups share the same id as the first ship group date fields.  The javascript renderDateField logic requires unique ids the all fields that are supposed to be created by renderDateField.

To reproduce:
1.  Create an order for DemoCustCompany
2.  Add 10 GZ-BASKET products to the order and click "Finalize Order"
3.  Add an additional ship group and click "Continue"
4.  Assign 5 of GZ-BASKET to second ship group and click "Continue"
5.  Complete the remaining steps required to complete the order 
6.  On the page after the "Create Order" link is clicked, notice the date fields of the second ship group were not rendered with a calendar.  

This patch allows all ship group date fields to be rendered correctly by changing the renderDateField id field from "shipBeforeDate" to "shipBeforeDate_${shipGroup.shipGroupSeqId}" and the same for shipAfterDate


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1539147 13f79535-47bb-0310-9956-ffa450edef68