A patch from Leon for "Form pagination fails if total length of search condition is more than about 2~3k bytes" https://issues.apache.org/jira/browse/OFBIZ-5478
Since it uses "GET" method to submit the pagination request, it's restricted by the URL max length limitation.
The patch attached changes the submit way from GET to "POST" if total length of URL > 2k.


jleroux: I also sort of (found after it was due to it) reverted r1204170, I don't see the reason for this now, and it was uselessly duplicating the current entry in the page size dropdown


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1559338 13f79535-47bb-0310-9956-ffa450edef68