A patch from Ritu Raj Lakhera  for "Authorize.net payment services are not handling  run time exceptions correctly" https://issues.apache.org/jira/browse/OFBIZ-5617
System is creating the order in CREATED status before processing the payment. But if payment processing encounters a run time exception then system throws the error and does not change the order status to REJECTED. Because Authorize.net payment services are not handling these exceptions correctly.

Ex: If authorize.net payment gateway call get interrupt then method processCard of AIMPaymentServices.java is returning the result map without the 'authResult' kay.
And subsequent method (processAuthTransResult) is trying to access this and thus service is throwing NullPointerException.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1588733 13f79535-47bb-0310-9956-ffa450edef68