A patch from Pierre Smits for "Be able to manually set preceding or succeeding production runs" https://issues.apache.org/jira/browse/OFBIZ-5523
In current feature set of the manufacturing compoment it is possible to set 'create dependent production run(s). This function will create the dependent production runs and have these shown in the production run.

In other scenario's of managing production runs it is not possible to associate one production run to another in the component. Although it might be needed/required. In order to do this now the user must (at least):
# have direct access to the workflow component (with appropriate permissions)
# look for the correct workeffort
# go to the children tab of the workeffort
# set the association

The improvement will be that the user of the manufacturing component will be able to define (in said component) the association between production runs.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1589869 13f79535-47bb-0310-9956-ffa450edef68