A patche from Olivier Heintz for "ProductStoreFacilities page does not refresh information after actions are performed" https://issues.apache.org/jira/browse/OFBIZ-5408
The ProductStoreFacilities page does not refresh when actions are performed.  When adding or deleting a facility from the store, the result of either action is not reflected on the page even though both are performed.  A loading icon appears after clicking action buttons/links so AJAX is probably being used.  

Olivier:
bug corrections :
1) areaId was wrong in on-event for edit and in delete button in list
2) replace wrong view-map ListProductStoreFacilityFormOnly by ListProductStoreFacility

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1598913 13f79535-47bb-0310-9956-ffa450edef68