Applied patch from jira issue - OFBIZ-5824 - Error in communication screen while perfom sorting. ========================================================================= In partymgr contrller.xml, we have view request named "ViewCommunicationEvent", and this request call an service event, due to this when we try to perform sorting on Child Communication Events, Customer Request List or any other screen over Communication screen then it will thorw following exception:
org.ofbiz.webapp.event.EventHandlerException: Found URL parameter [communicationEventId] passed to secure (https) request-map with uri [ViewCommunicationEvent] with an event that calls service [setCommEventRoleToRead]; this is not allowed for security reasons! The data should be encrypted by making it part of the request body (a form field) instead of the request URL. Moreover it would be kind if you could create a Jira sub-task of https://issues.apache.org/jira/browse/OFBIZ-2330 (check before if a sub-task for this error does not exist). If you are not sure how to create a Jira issue please have a look before at http://cwiki.apache.org/confluence/x/JIB2 Thank you in advance for your help.
 

Also due to this service event call on view request entry we can't open the communication events in new tab. Ideally this service should be call in screen context instead view request.

=====================================

Thanks Deepak for reporting the issue and providing the patch for the same. 



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1632745 13f79535-47bb-0310-9956-ffa450edef68