Applied patch from jira issue - OFBIZ-5783 - Shipment Box Type selection should be at package level instead of item level on Pack Order Screen.\ ======================================================================= The Shipment Box Type selection is at line item level on Pack Order screen where the package selection for each order item is made. It should instead be at package level next to the package weight input box.
Also, ShipmentBoxType values are supported only for USPS, which is hard coded in PackOrder.groovy.

There is a bug reported in OFBIZ-4606 for Release 11.04 but exists in Trunk as well. It should be fixed if we make the above change.
=======================================================================
Thanks Mridul for creating the issue and providing the patch for the same. 

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1632772 13f79535-47bb-0310-9956-ffa450edef68