Improve the way build-AREA_NAME targets are used - https://issues.apache.org/jira/browse/OFBIZ-5880
Since the introduction of the build-dev target, a build-AREA_NAME (test, qa, production) should not chain the main build target. Because servers use the patch command and patches applied by the build-dev target are rejected and the process fails.

Of course this also means to change the scripts which are using build-AREA_NAME since now the main build is not chained, eg
from
./ant clean build-qa
to 
./ant clean build-qa build

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1640196 13f79535-47bb-0310-9956-ffa450edef68